backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
PyMySQL
python-dotenv
requests